tests/inputdevice-script.html

<!DOCTYPE html>
<title>InputDevice test for script-generated objects</title>
<meta name="viewport" content="width=device-width">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevice-tests.js"></script>

<h1>InputDevice - test for script-generated objects</h1>
<div id="target"></div>
<div id="log"></div>

<script>
'use strict';
function runTests() {
  test(function() {
    var i = new InputDevice();
    assert_equals(i.firesTouchEvents, false, "firesTouchEvents exists and is false by default.");
    assert_readonly(i, "firesTouchEvents", "firesTouchEvents should be readonly");
  
    i = new InputDevice({});
    assert_equals(i.firesTouchEvents, false, "firesTouchEvents still false with empty init dictionary.");
  
    i = new InputDevice({firesTouchEvents:true});
    assert_equals(i.firesTouchEvents, true, "firesTouchEvents can be initialized to true.");
  
  }, "InputDevice objects can be created correctly from script");
  
  function generateConstructorTest(ctorName) {
    test(function() {
      var ctor = window[ctorName];
      var evt = new ctor("testEvent");
      assert_equals(evt.sourceDevice, null, "sourceDevice exists and is null by default");
      assert_readonly(evt, "sourceDevice", "sourceDevice should be readonly");
      
      var device = new InputDevice({firesTouchEvents:true});
      evt = new ctor("mytest", {sourceDevice: device});
      assert_equals(evt.sourceDevice, device, "sourceDevice can be initialized to a given InputDevice");
      
      var receivedEvent;
      var target = document.getElementById("target");
      target.addEventListener("mytest", function(e) {
        receivedEvent = e;
      });
      target.dispatchEvent(evt);
      assert_equals(receivedEvent.sourceDevice, device, "sourceDevice is preserved across JavaScript event dispatch");
      
    }, ctorName + ".sourceDevice exists and can be initialized properly by script");
  }

  var uiEventConstructors = ['UIEvent', 'MouseEvent', 'CompositionEvent', 'FocusEvent', 'KeyboardEvent', 'WheelEvent'];
  for (var i = 0; i < uiEventConstructors.length; i++)
    generateConstructorTest(uiEventConstructors[i]);

  // Note that SVGZoomEvent and TouchEvent don't have a constructor defined.
  var optionalUIEventConstructor = ['InputEvent', 'PointerEvent'];
  for (var i = 0; i < optionalUIEventConstructor.length; i++)
    if (optionalUIEventConstructor[i] in window)
      generateConstructorTest(optionalUIEventConstructor[i]);
}

</script>
